feature(server): 增加消息消费记录
增加记录，BaseHandler 添加后置处理保存消息记录表

+++ deliver-backend/deliver-backend-server/src/main/java/com/oszero/deliver/server/model/dto/request/SendRequestDto.java
@@ -24,7 +24,7 @@ public class SendRequestDto {
      * 消息模板 Id
      */
     @NotNull(message = "消息模板 ID 不能为 NULL")
-    private Integer templateId;
+    private Long templateId;
 
     /**
      * 发送用户列表
